⚡️ Crawler: allow floating-point `max_rps`

+++ src/opts.rs
@@ -211,8 +211,8 @@ pub struct ConnectionOpts {
     pub api_timeout: time::Duration,
 
     /// Maximum number of requests per second for the API.
-    #[clap(long, env = "BLITZ_DASHBOARD_MAX_API_RPS", default_value = "20")]
-    pub max_api_rps: u32,
+    #[clap(long, env = "BLITZ_DASHBOARD_MAX_API_RPS", default_value = "19.5")]
+    pub max_api_rps: f64,
 }
 
 #[derive(Parser)]
